feat: add RP init screen
