+++ ts/features/it-wallet/navigation/ItwRoutes.ts
@@ -20,6 +20,8 @@ export const ITW_ROUTES = {
   } as const,
   PRESENTATION: {
     PID_DETAILS: "ITW_PRESENTATION_PID_DETAILS",
-    PID_ATTRIBUTES: "ITW_PRESENTATION_PID_ATTRIBUTES"
+    CROSS_DEVICE: {
+      INIT: "ITW_PRESENTATION_CROSS_DEVICE_INIT"
+    } as const
   }
 } as const;
