+++ ts/features/it-wallet/saga/itwRpSaga.ts
@@ -1,7 +1,8 @@
 import { SagaIterator } from "redux-saga";
-import { call, put, take, takeLatest } from "typed-redux-saga/macro";
+import { call, put, select, take, takeLatest } from "typed-redux-saga/macro";
 import { ActionType, getType } from "typesafe-actions";
 import { RelyingPartySolution } from "@pagopa/io-react-native-wallet";
+import { CommonActions } from "@react-navigation/native";
 import {
   itwRpCompleted,
   itwRpInitialization,
@@ -12,8 +13,12 @@ import {
   itwRpUserRejected
 } from "../store/actions/itwRpActions";
 import { itwWiaRequest } from "../store/actions/itwWiaActions";
-import { handleItwRpInitializationSaga } from "./itwRpInitialization";
+import { itwDecodePid } from "../store/actions/itwCredentialsActions";
+import { itwPidValueSelector } from "../store/reducers/itwPidReducer";
+import NavigationService from "../../../navigation/NavigationService";
+import ROUTES from "../../../navigation/routes";
 import { handleItwRpPresentationSaga } from "./itwRpPresentation";
+import { handleItwRpInitializationSaga } from "./itwRpInitialization";
 
 /**
  * Watcher for the IT wallet Relying Party related sagas.
@@ -54,6 +59,16 @@ export function* handleRpStart(
   yield* call(itwWiaRequest.request);
   const wia = yield* take(itwWiaRequest.success);
 
+  // Decode PID
+  const pid = yield* select(itwPidValueSelector);
+  yield* call(itwDecodePid.request, pid);
+  const decodedRes = yield* take<
+    ActionType<typeof itwWiaRequest.success | typeof itwWiaRequest.failure>
+  >([itwWiaRequest.success, itwWiaRequest.failure]);
+  if (decodedRes.type === getType(itwWiaRequest.failure)) {
+    yield* put(itwRpStop());
+  }
+
   // The RP solution is initialized using the authReqUrl
   // of the qrcode payload
   const RP = new RelyingPartySolution(authReqUrl, wia.payload);
@@ -68,8 +83,16 @@ export function* handleRpStart(
   }
 }
 
+/**
+ * On RP stop, the user is redirected to the IT wallet home.
+ */
 export function* handleRpStop(): SagaIterator {
-  // TODO: (ex. go back or reset stack navigation)
+  yield* call(
+    NavigationService.dispatchNavigationAction,
+    CommonActions.navigate(ROUTES.WALLET_NAVIGATOR, {
+      screen: ROUTES.ITWALLET_HOME
+    })
+  );
 }
 
 export function* handleRpCompleted(): SagaIterator {
